requirements.txt
openai==1.91.0
agents==1.4.0
requests
python-dotenv
